Split FFI module out into 2 submodules
